+++ extension/src/components/VotingControls.vue
@@ -36,150 +36,37 @@
       </button>
     </div>
 
-    <!-- Status Change Modal -->
-    <div v-if="showStatusModal" class="modal-overlay" @click="closeStatusModal">
-      <div class="status-modal" @click.stop>
-        <div class="modal-header">
-          <h3>Change Status</h3>
-          <button class="close-btn" @click="closeStatusModal">×</button>
-        </div>
-        
-        <div class="modal-content">
-          <p><strong>Proposal:</strong> #{{ proposalId }}</p>
-          <p><strong>Current Status:</strong> {{ status }}</p>
-          
-          <div class="status-options">
-            <label>New Status:</label>
-            <div class="status-grid">
-              <button
-                v-for="statusOption in statusOptions"
-                :key="statusOption.value"
-                class="status-option"
-                :class="{ selected: selectedStatus === statusOption.value }"
-                @click="selectedStatus = statusOption.value"
-              >
-                <span class="option-icon">{{ statusOption.icon }}</span>
-                <span class="option-text">{{ statusOption.value }}</span>
-              </button>
-            </div>
-          </div>
-          
-          <div class="reason-section">
-            <label for="reason">Reason for change (optional):</label>
-            <textarea
-              id="reason"
-              v-model="changeReason"
-              placeholder="Explain why you're changing the status..."
-              rows="3"
-            ></textarea>
-          </div>
-          
-          <div class="modal-actions">
-            <button class="btn btn-secondary" @click="closeStatusModal">Cancel</button>
-            <button 
-              class="btn btn-primary" 
-              @click="saveStatusChange"
-              :disabled="!selectedStatus || selectedStatus === status"
-            >
-              Update Status
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <!-- Modals -->
+    <StatusChangeModal 
+      :show="showStatusModal"
+      :proposal-id="proposalId"
+      :current-status="status"
+      @close="closeStatusModal"
+      @save="saveStatusChange"
+    />
 
-    <!-- Assign Modal -->
-    <div v-if="showAssignModal" class="modal-overlay" @click="closeAssignModal">
-      <div class="assign-modal" @click.stop>
-        <div class="modal-header">
-          <h3>Assign Proposal</h3>
-          <button class="close-btn" @click="closeAssignModal">×</button>
-        </div>
-        
-        <div class="modal-content">
-          <p><strong>Proposal:</strong> #{{ proposalId }}</p>
-          <p>This will assign the proposal to you for review and voting.</p>
-          
-          <div class="modal-actions">
-            <button class="btn btn-secondary" @click="closeAssignModal">Cancel</button>
-            <button class="btn btn-primary" @click="confirmAssign">
-              Assign to Me
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <AssignModal 
+      :show="showAssignModal"
+      :proposal-id="proposalId"
+      @close="closeAssignModal"
+      @confirm="confirmAssign"
+    />
 
-    <!-- Vote Change Modal -->
-    <div v-if="showVoteModal" class="modal-overlay" @click="closeVoteModal">
-      <div class="vote-modal" @click.stop>
-        <div class="modal-header">
-          <h3>Change Suggested Vote</h3>
-          <button class="close-btn" @click="closeVoteModal">×</button>
-        </div>
-        
-        <div class="modal-content">
-          <p><strong>Proposal:</strong> #{{ proposalId }}</p>
-          <p>Select your suggested vote for this proposal:</p>
-          
-          <div class="vote-options">
-            <button
-              class="vote-option aye"
-              :class="{ selected: selectedVote === 'aye' }"
-              @click="selectedVote = 'aye'"
-            >
-              <span class="vote-icon">👍</span>
-              <span class="vote-text">Aye</span>
-            </button>
-            
-            <button
-              class="vote-option nay"
-              :class="{ selected: selectedVote === 'nay' }"
-              @click="selectedVote = 'nay'"
-            >
-              <span class="vote-icon">👎</span>
-              <span class="vote-text">Nay</span>
-            </button>
-            
-            <button
-              class="vote-option abstain"
-              :class="{ selected: selectedVote === 'abstain' }"
-              @click="selectedVote = 'abstain'"
-            >
-              <span class="vote-icon">✌️</span>
-              <span class="vote-text">Abstain</span>
-            </button>
-          </div>
-          
-          <div class="reason-section">
-            <label for="vote-reason">Reason (optional):</label>
-            <textarea
-              id="vote-reason"
-              v-model="voteReason"
-              placeholder="Explain your vote suggestion..."
-              rows="3"
-            ></textarea>
-          </div>
-          
-          <div class="modal-actions">
-            <button class="btn btn-secondary" @click="closeVoteModal">Cancel</button>
-            <button 
-              class="btn btn-primary" 
-              @click="saveVoteChange"
-              :disabled="!selectedVote"
-            >
-              Update Vote
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <VoteChangeModal 
+      :show="showVoteModal"
+      :proposal-id="proposalId"
+      @close="closeVoteModal"
+      @save="saveVoteChange"
+    />
   </div>
 </template>
 
 <script setup lang="ts">
 import { ref, computed } from 'vue'
 import type { InternalStatus, SuggestedVote } from '../types'
+import StatusChangeModal from './StatusChangeModal.vue'
+import AssignModal from './AssignModal.vue'
+import VoteChangeModal from './VoteChangeModal.vue'
 
 interface VotingControlsProps {
   status: InternalStatus
@@ -191,18 +78,10 @@ interface VotingControlsProps {
 
 const props = defineProps<VotingControlsProps>()
 
-// Status modal state
+// Modal states
 const showStatusModal = ref(false)
-const selectedStatus = ref<InternalStatus>(props.status)
-const changeReason = ref('')
-
-// Assign modal state
 const showAssignModal = ref(false)
-
-// Vote modal state
 const showVoteModal = ref(false)
-const selectedVote = ref<'aye' | 'nay' | 'abstain' | null>(null)
-const voteReason = ref('')
 
 const statusConfig = {
   'Not started': { color: '#6c757d', icon: '⚪' },
@@ -217,12 +96,6 @@ const statusConfig = {
   'Not Voted': { color: '#e9ecef', icon: '❌' }
 }
 
-const statusOptions = Object.keys(statusConfig).map(status => ({
-  value: status as InternalStatus,
-  icon: statusConfig[status as InternalStatus].icon,
-  color: statusConfig[status as InternalStatus].color
-}))
-
 const statusClass = computed(() => {
   const config = statusConfig[props.status]
   return {
@@ -236,26 +109,20 @@ const statusIcon = computed(() => statusConfig[props.status]?.icon || '⚪')
 const handleStatusClick = () => {
   if (props.editable) {
     showStatusModal.value = true
-    selectedStatus.value = props.status
-    changeReason.value = ''
   }
 }
 
 const closeStatusModal = () => {
   showStatusModal.value = false
-  selectedStatus.value = props.status
-  changeReason.value = ''
 }
 
-const saveStatusChange = async () => {
-  if (!selectedStatus.value || selectedStatus.value === props.status) return
-  
+const saveStatusChange = async (data: { newStatus: InternalStatus; reason: string }) => {
   try {
     const changeData = {
       proposalId: props.proposalId,
       oldStatus: props.status,
-      newStatus: selectedStatus.value,
-      reason: changeReason.value
+      newStatus: data.newStatus,
+      reason: data.reason
     }
     
     console.log('Status change requested:', changeData)
@@ -299,24 +166,18 @@ const confirmAssign = async () => {
 const handleChangeVote = () => {
   if (!props.isAuthenticated) return
   showVoteModal.value = true
-  selectedVote.value = null
-  voteReason.value = ''
 }
 
 const closeVoteModal = () => {
   showVoteModal.value = false
-  selectedVote.value = null
-  voteReason.value = ''
 }
 
-const saveVoteChange = async () => {
-  if (!selectedVote.value) return
-  
+const saveVoteChange = async (data: { vote: 'aye' | 'nay' | 'abstain'; reason: string }) => {
   try {
     const voteData = {
       proposalId: props.proposalId,
-      vote: selectedVote.value,
-      reason: voteReason.value
+      vote: data.vote,
+      reason: data.reason
     }
     
     console.log('Vote change requested:', voteData)
@@ -501,264 +362,7 @@ const saveVoteChange = async () => {
   font-size: 0.9rem;
 }
 
-/* Modal styles */
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000000;
-  backdrop-filter: blur(2px);
-}
-
-.status-modal,
-.assign-modal,
-.vote-modal {
-  background: white;
-  border-radius: 12px;
-  width: 90%;
-  max-width: 500px;
-  max-height: 80vh;
-  overflow-y: auto;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.2);
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px 24px;
-  border-bottom: 1px solid #e9ecef;
-  background: linear-gradient(135deg, #e6007a, #b3005f);
-  color: white;
-  border-radius: 12px 12px 0 0;
-}
-
-.modal-header h3 {
-  margin: 0;
-  font-size: 1.2rem;
-  font-weight: 600;
-}
-
-.close-btn {
-  background: rgba(255, 255, 255, 0.2);
-  border: none;
-  color: white;
-  font-size: 1.5rem;
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: background-color 0.2s ease;
-}
-
-.close-btn:hover {
-  background: rgba(255, 255, 255, 0.3);
-}
-
-.modal-content {
-  padding: 24px;
-}
-
-.modal-content p {
-  margin: 0 0 16px 0;
-  color: #495057;
-}
-
-.status-options,
-.vote-options {
-  margin: 16px 0;
-}
-
-.status-options label {
-  display: block;
-  margin-bottom: 8px;
-  font-weight: 600;
-  color: #495057;
-}
-
-.status-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(120px, 1fr));
-  gap: 8px;
-  margin-bottom: 16px;
-}
-
-.status-option {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 4px;
-  padding: 12px 8px;
-  border: 2px solid #e9ecef;
-  border-radius: 8px;
-  background: white;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  font-size: 0.8rem;
-}
-
-.status-option:hover {
-  border-color: #e6007a;
-  box-shadow: 0 2px 8px rgba(230, 0, 122, 0.15);
-}
-
-.status-option.selected {
-  border-color: #e6007a;
-  background: linear-gradient(135deg, #fff5f8, #ffe8f0);
-}
-
-.option-icon {
-  font-size: 1.2rem;
-}
-
-.option-text {
-  text-align: center;
-  font-weight: 500;
-}
-
-.vote-options {
-  display: flex;
-  gap: 12px;
-  margin: 16px 0;
-}
-
-.vote-option {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  padding: 16px 12px;
-  border: 2px solid #e9ecef;
-  border-radius: 8px;
-  background: white;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  font-size: 0.9rem;
-}
-
-.vote-option:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-}
-
-.vote-option.selected {
-  border-color: #e6007a;
-  background: linear-gradient(135deg, #fff5f8, #ffe8f0);
-}
-
-.vote-option.aye:hover,
-.vote-option.aye.selected {
-  border-color: #28a745;
-  background: linear-gradient(135deg, #f8fff9, #e8f5e8);
-}
-
-.vote-option.nay:hover,
-.vote-option.nay.selected {
-  border-color: #dc3545;
-  background: linear-gradient(135deg, #fff8f8, #ffe8e8);
-}
-
-.vote-option.abstain:hover,
-.vote-option.abstain.selected {
-  border-color: #6f42c1;
-  background: linear-gradient(135deg, #faf8ff, #f0e8ff);
-}
-
-.vote-icon {
-  font-size: 1.5rem;
-}
-
-.vote-text {
-  font-weight: 600;
-}
-
-.reason-section {
-  margin: 16px 0;
-}
-
-.reason-section label {
-  display: block;
-  margin-bottom: 8px;
-  font-weight: 600;
-  color: #495057;
-}
-
-.reason-section textarea {
-  width: 100%;
-  padding: 12px;
-  border: 1px solid #e9ecef;
-  border-radius: 8px;
-  font-family: inherit;
-  font-size: 0.9rem;
-  resize: vertical;
-  transition: border-color 0.2s ease;
-}
-
-.reason-section textarea:focus {
-  outline: none;
-  border-color: #e6007a;
-  box-shadow: 0 0 0 3px rgba(230, 0, 122, 0.1);
-}
-
-.modal-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 12px;
-  margin-top: 24px;
-  padding-top: 16px;
-  border-top: 1px solid #e9ecef;
-}
-
-.btn {
-  padding: 10px 20px;
-  border: none;
-  border-radius: 8px;
-  font-size: 0.9rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  text-decoration: none;
-  white-space: nowrap;
-}
-
-.btn:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-.btn-primary {
-  background: linear-gradient(135deg, #e6007a, #b3005f);
-  color: white;
-  border: 1px solid #b3005f;
-}
-
-.btn-primary:hover:not(:disabled) {
-  background: linear-gradient(135deg, #b3005f, #8a0047);
-  transform: translateY(-1px);
-  box-shadow: 0 4px 12px rgba(230, 0, 122, 0.3);
-}
-
-.btn-secondary {
-  background: #6c757d;
-  color: white;
-  border: 1px solid #6c757d;
-}
-
-.btn-secondary:hover:not(:disabled) {
-  background: #5a6268;
-  transform: translateY(-1px);
-  box-shadow: 0 4px 12px rgba(108, 117, 125, 0.3);
-}
+/* Removed modal styles - now in separate modal components */
 
 @media (max-width: 768px) {
   .controls-actions {
